automatically backed up by learn

+++ index.js
@@ -20,7 +20,8 @@ function makeInt(n) {
   return n
 }
 function preserveDecimal(n) {
-  return parseInt()
+  return parseFloat()
 }
+
